+++ pystac_client/client.py
@@ -1,14 +1,17 @@
 from functools import lru_cache
+import re
 from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
+    cast,
     Dict,
     Iterator,
     List,
     Optional,
     Union,
 )
+import warnings
 
 import pystac
 import pystac.utils
@@ -19,6 +22,7 @@ from requests import Request
 from pystac_client._utils import Modifiable, call_modifier
 from pystac_client.collection_client import CollectionClient
 from pystac_client.conformance import ConformanceClasses
+
 from pystac_client.errors import ClientTypeError
 from pystac_client.exceptions import APIError
 from pystac_client.item_search import (
@@ -37,6 +41,11 @@ from pystac_client.item_search import (
 )
 from pystac_client.mixins import QueryablesMixin
 from pystac_client.stac_api_io import StacApiIO, Timeout
+from pystac_client.warnings import (
+    DoesNotConformTo,
+    FallbackToPystac,
+    NoConformsTo,
+)
 
 if TYPE_CHECKING:
     from pystac.item import Item as Item_Type
@@ -93,7 +102,7 @@ class Client(pystac.Catalog, QueryablesMixin):
         url: str,
         headers: Optional[Dict[str, str]] = None,
         parameters: Optional[Dict[str, Any]] = None,
-        ignore_conformance: bool = False,
+        ignore_conformance: Optional[bool] = None,
         modifier: Optional[Callable[[Modifiable], None]] = None,
         request_modifier: Optional[Callable[[Request], Union[Request, None]]] = None,
         stac_io: Optional[StacApiIO] = None,
@@ -108,11 +117,16 @@ class Client(pystac.Catalog, QueryablesMixin):
                 made to any part of this Catalog/API.
             parameters: Optional dictionary of query string parameters to
                 include in all requests.
-            ignore_conformance : Ignore any advertised Conformance Classes in this
-                Catalog/API. This means that
+            ignore_conformance (DEPRECATED) : Ignore any advertised Conformance Classes
+                in this Catalog/API. This means that
                 functions will skip checking conformance, and may throw an unknown
                 error if that feature is
                 not supported, rather than a :class:`NotImplementedError`.
+
+                .. deprecated:: 0.7.0
+                    Conformance can be altered rather than ignored using methods like
+                    :meth:`clear_conforms_to` and :meth:`add_conforms_to`
+
             modifier : A callable that modifies the children collection and items
                 returned by this Client. This can be useful for injecting
                 authentication parameters into child assets to access data
@@ -163,21 +177,19 @@ class Client(pystac.Catalog, QueryablesMixin):
             stac_io=stac_io,
             timeout=timeout,
         )
-        search_link = client.get_search_link()
-        # if there is a search link, but no conformsTo advertised, ignore
-        # conformance entirely
-        # NOTE: this behavior to be deprecated as implementations become conformant
-        if client._stac_io and (
-            ignore_conformance
-            or (
-                client
-                and "conformsTo" not in client.extra_fields.keys()
-                and search_link
-                and search_link.href
-                and len(search_link.href) > 0
+
+        if ignore_conformance is not None:
+            warnings.warn(
+                (
+                    "The `ignore_conformance` option is deprecated and will be "
+                    "removed in the next major release. Instead use `set_conforms_to` "
+                    "or `add_conforms_to` to control behavior."
+                ),
+                FutureWarning,
             )
-        ):
-            client._stac_io.set_conformance(None)
+
+        if not client.has_conforms_to():
+            warnings.warn(NoConformsTo())
 
         return client
 
@@ -213,21 +225,86 @@ class Client(pystac.Catalog, QueryablesMixin):
             )
 
         client: Client = super().from_file(href, stac_io)
-
-        client._stac_io._conformance = client.extra_fields.get(  # type: ignore
-            "conformsTo", []
-        )
         client.modifier = modifier
 
         return client
 
-    def _supports_collections(self) -> bool:
-        return self._conforms_to(ConformanceClasses.COLLECTIONS) or self._conforms_to(
-            ConformanceClasses.FEATURES
+    def has_conforms_to(self) -> bool:
+        """Whether server contains list of ``"conformsTo"`` URIs"""
+        return "conformsTo" in self.extra_fields
+
+    def get_conforms_to(self) -> List[str]:
+        """List of ``"conformsTo"`` URIs
+
+        Return:
+            List[str]: List of  URIs that the server conforms to
+        """
+        return cast(List[str], self.extra_fields.get("conformsTo", []).copy())
+
+    def set_conforms_to(self, conformance_uris: List[str]) -> None:
+        """Set list of ``"conformsTo"`` URIs
+
+        Args:
+            conformance_uris : URIs indicating what the server conforms to
+        """
+        self.extra_fields["conformsTo"] = conformance_uris
+
+    def clear_conforms_to(self) -> None:
+        """Clear list of ``"conformsTo"`` urls
+
+        Removes the entire list, so :py:meth:`has_conforms_to` will
+        return False after using this method.
+        """
+        self.extra_fields.pop("conformsTo", None)
+
+    def add_conforms_to(self, name: str) -> None:
+        """Add ``"conformsTo"`` by name.
+
+        Args:
+            name : name of :py:class:`ConformanceClasses` keys to add.
+        """
+        conformance_class = ConformanceClasses.get_by_name(name)
+
+        if not self.conforms_to(conformance_class):
+            self.set_conforms_to([*self.get_conforms_to(), conformance_class.valid_uri])
+
+    def remove_conforms_to(self, name: str) -> None:
+        """Remove ``"conformsTo"`` by name.
+
+        Args:
+            name : name of :py:class:`ConformanceClasses` keys to remove.
+        """
+        conformance_class = ConformanceClasses.get_by_name(name)
+
+        self.set_conforms_to(
+            [
+                uri
+                for uri in self.get_conforms_to()
+                if not re.match(conformance_class.pattern, uri)
+            ]
         )
 
-    def _conforms_to(self, conformance_class: ConformanceClasses) -> bool:
-        return self._stac_io.conforms_to(conformance_class)  # type: ignore
+    def conforms_to(self, conformance_class: Union[ConformanceClasses, str]) -> bool:
+        """Checks whether the API conforms to the given standard.
+
+        This method only checks
+        against the ``"conformsTo"`` property from the API landing page and does not
+        make any additional calls to a ``/conformance`` endpoint even if the API
+        provides such an endpoint.
+
+        Args:
+            name : name of :py:class:`ConformanceClasses` keys to check
+                conformance against.
+
+        Return:
+            bool: Indicates if the API conforms to the given spec or URI.
+        """
+        if isinstance(conformance_class, str):
+            conformance_class = ConformanceClasses.get_by_name(conformance_class)
+
+        return any(
+            re.match(conformance_class.pattern, uri) for uri in self.get_conforms_to()
+        )
 
     @classmethod
     def from_dict(
@@ -253,6 +330,16 @@ class Client(pystac.Catalog, QueryablesMixin):
         result.modifier = modifier
         return result
 
+    def _supports_collections(self) -> bool:
+        return self.conforms_to(ConformanceClasses.COLLECTIONS) or self.conforms_to(
+            ConformanceClasses.FEATURES
+        )
+
+    def _warn_about_fallback(self, *args: str) -> None:
+        if self.has_conforms_to():
+            warnings.warn(DoesNotConformTo(*args), stacklevel=2)
+        warnings.warn(FallbackToPystac(), stacklevel=2)
+
     @lru_cache()
     def get_collection(
         self, collection_id: str
@@ -263,38 +350,42 @@ class Client(pystac.Catalog, QueryablesMixin):
             collection_id: The Collection ID to get
 
         Returns:
-            CollectionClient: A STAC Collection
+            Union[Collection, CollectionClient]: A STAC Collection
         """
-        if self._supports_collections() and self._stac_io:
-            url = self._get_collections_href(collection_id)
+        collection: Union[Collection, CollectionClient]
+
+        if self._supports_collections():
+            assert self._stac_io is not None
+
+            url = self._collections_href(collection_id)
             collection = CollectionClient.from_dict(
                 self._stac_io.read_json(url),
                 root=self,
                 modifier=self.modifier,
             )
             call_modifier(self.modifier, collection)
-            return collection
         else:
-            for col in self.get_collections():
-                if col.id == collection_id:
-                    call_modifier(self.modifier, col)
-                    return col
+            self._warn_about_fallback("COLLECTIONS", "FEATURES")
+            for collection in super().get_collections():
+                if collection.id == collection_id:
+                    call_modifier(self.modifier, collection)
+        return collection
 
-        return None
-
-    def get_collections(self) -> Iterator[Union[Collection, CollectionClient]]:
+    def get_collections(self) -> Iterator[Collection]:
         """Get Collections in this Catalog
 
             Gets the collections from the /collections endpoint if supported,
             otherwise fall back to Catalog behavior of following child links
 
         Return:
-            Iterator[Collection]: Iterator over Collections in Catalog/API
+            Iterator[Union[Collection, CollectionClient]]: Collections in Catalog/API
         """
         collection: Union[Collection, CollectionClient]
 
-        if self._supports_collections() and self._stac_io:
-            url = self._get_collections_href()
+        if self._supports_collections():
+            assert self._stac_io is not None
+
+            url = self._collections_href()
             for page in self._stac_io.get_pages(url):
                 if "collections" not in page:
                     raise APIError("Invalid response from /collections")
@@ -305,6 +396,7 @@ class Client(pystac.Catalog, QueryablesMixin):
                     call_modifier(self.modifier, collection)
                     yield collection
         else:
+            self._warn_about_fallback("COLLECTIONS", "FEATURES")
             for collection in super().get_collections():
                 call_modifier(self.modifier, collection)
                 yield collection
@@ -316,10 +408,11 @@ class Client(pystac.Catalog, QueryablesMixin):
             Iterator[Item]:: Iterator of items whose parent is this
                 catalog.
         """
-        if self._conforms_to(ConformanceClasses.ITEM_SEARCH):
+        if self.conforms_to(ConformanceClasses.ITEM_SEARCH):
             search = self.search()
             yield from search.items()
         else:
+            self._warn_about_fallback("ITEM_SEARCH")
             for item in super().get_items():
                 call_modifier(self.modifier, item)
                 yield item
@@ -333,13 +426,7 @@ class Client(pystac.Catalog, QueryablesMixin):
                 catalogs or collections connected to this catalog through
                 child links.
         """
-        if self._conforms_to(ConformanceClasses.ITEM_SEARCH):
-            # these are already modified
-            yield from self.get_items()
-        else:
-            for item in super().get_items():
-                call_modifier(self.modifier, item)
-                yield item
+        yield from self.get_items()
 
     def search(
         self,
@@ -458,29 +545,16 @@ class Client(pystac.Catalog, QueryablesMixin):
                 or does not have a link with
                 a ``"rel"`` type of ``"search"``.
         """
-        if not self._conforms_to(ConformanceClasses.ITEM_SEARCH):
-            raise NotImplementedError(
-                "This catalog does not support search because it "
-                f'does not conform to "{ConformanceClasses.ITEM_SEARCH}"'
-            )
-        search_link = self.get_search_link()
-        if search_link:
-            if isinstance(search_link.target, str):
-                search_href = search_link.target
-            else:
-                raise NotImplementedError(
-                    "Link with rel=search was an object rather than a URI"
-                )
-        else:
-            raise NotImplementedError(
-                "No link with rel=search could be found in this catalog"
+
+        if not self.conforms_to(ConformanceClasses.ITEM_SEARCH):
+            raise DoesNotConformTo(
+                "ITEM_SEARCH", "There is not fallback option available for search."
             )
 
         return ItemSearch(
-            url=search_href,
+            url=self._search_href(),
             method=method,
             max_items=max_items,
-            stac_io=self._stac_io,
             client=self,
             limit=limit,
             ids=ids,
@@ -517,35 +591,14 @@ class Client(pystac.Catalog, QueryablesMixin):
             None,
         )
 
-    def _get_collections_href(self, collection_id: Optional[str] = None) -> str:
-        self_href = self.get_self_href()
-        if self_href is None:
-            data_link = self.get_single_link("data")
-            if data_link is None:
-                raise ValueError(
-                    "cannot build a collections href without a self href or a data link"
-                )
-            else:
-                collections_href = data_link.href
-        else:
-            collections_href = f"{self_href.rstrip('/')}/collections"
+    def _search_href(self) -> str:
+        search_link = self.get_search_link()
+        href = self._get_href("search", search_link, "search")
+        return href
 
-        if not pystac.utils.is_absolute_href(collections_href):
-            collections_href = self._make_absolute_href(collections_href)
-
-        if collection_id is None:
-            return collections_href
-        else:
-            return f"{collections_href.rstrip('/')}/{collection_id}"
-
-    def _make_absolute_href(self, href: str) -> str:
-        self_link = self.get_single_link("self")
-        if self_link is None:
-            raise ValueError("cannot build an absolute href without a self link")
-        elif not pystac.utils.is_absolute_href(self_link.href):
-            raise ValueError(
-                "cannot build an absolute href from "
-                f"a relative self link: {self_link.href}"
-            )
-        else:
-            return pystac.utils.make_absolute_href(href, self_link.href)
+    def _collections_href(self, collection_id: Optional[str] = None) -> str:
+        data_link = self.get_single_link("data")
+        href = self._get_href("data", data_link, "collections")
+        if collection_id is not None:
+            return f"{href.rstrip('/')}/{collection_id}"
+        return href