+++ pystac_client/conformance.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import re
 from enum import Enum
 
@@ -5,22 +7,42 @@ from enum import Enum
 class ConformanceClasses(Enum):
     """Enumeration class for Conformance Classes"""
 
-    stac_prefix = re.escape("https://api.stacspec.org/v1.0.")
-
     # defined conformance classes regexes
-    CORE = rf"{stac_prefix}(.*){re.escape('/core')}"
-    COLLECTIONS = rf"{stac_prefix}(.*){re.escape('/collections')}"
+    CORE = "/core"
+    COLLECTIONS = "/collections"
 
     # this is ogcapi-features instead of just features for historical reasons,
     # even thought this is a STAC API conformance class
-    FEATURES = rf"{stac_prefix}(.*){re.escape('/ogcapi-features')}"
-    ITEM_SEARCH = rf"{stac_prefix}(.*){re.escape('/item-search')}"
+    FEATURES = "/ogcapi-features"
+    ITEM_SEARCH = "/item-search"
 
-    CONTEXT = rf"{stac_prefix}(.*){re.escape('/item-search#context')}"
-    FIELDS = rf"{stac_prefix}(.*){re.escape('/item-search#fields')}"
-    SORT = rf"{stac_prefix}(.*){re.escape('/item-search#sort')}"
-    QUERY = rf"{stac_prefix}(.*){re.escape('/item-search#query')}"
-    FILTER = rf"{stac_prefix}(.*){re.escape('/item-search#filter')}"
+    CONTEXT = "/item-search#context"
+    FIELDS = "/item-search#fields"
+    SORT = "/item-search#sort"
+    QUERY = "/item-search#query"
+    FILTER = "/item-search#filter"
 
+    @classmethod
+    def get_by_name(cls, name: str) -> ConformanceClasses:
+        for member in cls:
+            if member.name == name.upper():
+                return member
+        raise ValueError(
+            f"Invalid conformance class '{name}'. Options are: {list(cls)}"
+        )
 
-CONFORMANCE_URIS = {c.name: c.value for c in ConformanceClasses}
+    def __str__(self) -> str:
+        return f"{self.name}"
+
+    def __repr__(self) -> str:
+        return str(self)
+
+    @property
+    def valid_uri(self) -> str:
+        return f"https://api.stacspec.org/v1.0.*{self.value}"
+
+    @property
+    def pattern(self) -> re.Pattern[str]:
+        return re.compile(
+            rf"{re.escape('https://api.stacspec.org/v1.0.')}(.*){re.escape(self.value)}"
+        )