+++ pystac_client/mixins.py
@@ -1,20 +1,35 @@
-from typing import Optional, Dict, Any
+from typing import Optional, Dict, Any, Union
+import warnings
 
 import pystac
 
 from pystac_client.exceptions import APIError
 from pystac_client.conformance import ConformanceClasses
 from pystac_client.stac_api_io import StacApiIO
+from pystac_client.warnings import DoesNotConformTo, MissingLink
 
 QUERYABLES_REL = "http://www.opengis.net/def/rel/ogc/1.0/queryables"
-QUERYABLES_ENDPOINT = "/queryables"
+QUERYABLES_ENDPOINT = "queryables"
 
 
 class StacAPIObject(pystac.STACObject):
     _stac_io: Optional[StacApiIO]
 
+    def conforms_to(self, conformance_class: Union[str, ConformanceClasses]) -> bool:
+        raise NotImplementedError
 
-class QueryablesMixin(StacAPIObject):
+
+class BaseMixin(StacAPIObject):
+    def _get_href(self, rel: str, link: Optional[pystac.Link], endpoint: str) -> str:
+        if link and isinstance(link.href, str):
+            href = link.absolute_href
+        else:
+            warnings.warn(MissingLink(rel, self.__class__.__name__), stacklevel=2)
+            href = f"{self.self_href.rstrip('/')}/{endpoint}"
+        return href
+
+
+class QueryablesMixin(BaseMixin):
     """Mixin for adding support for /queryables endpoint"""
 
     def get_queryables(self) -> Dict[str, Any]:
@@ -28,7 +43,9 @@ class QueryablesMixin(StacAPIObject):
         if self._stac_io is None:
             raise APIError("API access is not properly configured")
 
-        self._stac_io.assert_conforms_to(ConformanceClasses.FILTER)
+        if not self.conforms_to(ConformanceClasses.FILTER):
+            raise DoesNotConformTo(ConformanceClasses.FILTER.name)
+
         url = self._get_queryables_href()
 
         result = self._stac_io.read_json(url)
@@ -42,10 +59,5 @@ class QueryablesMixin(StacAPIObject):
 
     def _get_queryables_href(self) -> str:
         link = self.get_single_link(QUERYABLES_REL)
-        if link is not None:
-            url = link.href
-        else:
-            # The queryables link should be defined at the root, but if it is not
-            # try to guess the url
-            url = f"{self.self_href.rstrip('/')}{QUERYABLES_ENDPOINT}"
-        return url
+        href = self._get_href(QUERYABLES_REL, link, QUERYABLES_ENDPOINT)
+        return href