+++ pystac_client/stac_api_io.py
@@ -1,6 +1,5 @@
 import json
 import logging
-import re
 from copy import deepcopy
 from typing import (
     TYPE_CHECKING,
@@ -15,6 +14,7 @@ from typing import (
 )
 from typing_extensions import TypeAlias
 from urllib.parse import urlparse
+import warnings
 
 import pystac
 from pystac.link import Link
@@ -29,7 +29,6 @@ from requests import Request, Session
 
 import pystac_client
 
-from .conformance import CONFORMANCE_URIS, ConformanceClasses
 from .exceptions import APIError
 
 if TYPE_CHECKING:
@@ -55,8 +54,12 @@ class StacApiIO(DefaultStacIO):
 
         Args:
             headers : Optional dictionary of headers to include in all requests
-            conformance : Optional list of `Conformance Classes
+            conformance (DEPRECATED) : Optional list of `Conformance Classes
                 <https://github.com/radiantearth/stac-api-spec/blob/master/overview.md#conformance-classes>`__.
+
+                .. deprecated:: 0.7.0
+                    Conformance can be altered on the client class directly
+
             parameters: Optional dictionary of query string parameters to
               include in all requests.
             request_modifier: Optional callable that can be used to modify Request
@@ -71,8 +74,19 @@ class StacApiIO(DefaultStacIO):
             StacApiIO : StacApiIO instance
         """
         # TODO - this should super() to parent class
+
+        if conformance is not None:
+            warnings.warn(
+                (
+                    "The `conformance` option is deprecated and will be "
+                    "removed in the next major release. Instead use "
+                    "`Client.set_conforms_to` or `Client.add_conforms_to` to control "
+                    "behavior."
+                ),
+                category=FutureWarning,
+            )
+
         self.session = Session()
-        self._conformance = conformance
         self.timeout = timeout
         self.update(
             headers=headers, parameters=parameters, request_modifier=request_modifier
@@ -247,6 +261,7 @@ class StacApiIO(DefaultStacIO):
             return result
 
         if info.object_type == pystac.STACObjectType.COLLECTION:
+            assert isinstance(root, pystac_client.client.Client)
             return pystac_client.collection_client.CollectionClient.from_dict(
                 d, href=str(href), root=root, migrate=False, preserve_dict=preserve_dict
             )
@@ -289,54 +304,6 @@ class StacApiIO(DefaultStacIO):
             next_link = next(
                 (link for link in page.get("links", []) if link["rel"] == "next"), None
             )
-
-    def assert_conforms_to(self, conformance_class: ConformanceClasses) -> None:
-        """Raises a :exc:`NotImplementedError` if the API does not publish the given
-        conformance class. This method only checks against the ``"conformsTo"``
-        property from the API landing page and does not make any additional
-        calls to a ``/conformance`` endpoint even if the API provides such an endpoint.
-
-        Args:
-            conformance_class: The ``ConformanceClasses`` key to check conformance
-            against.
-        """
-        if not self.conforms_to(conformance_class):
-            raise NotImplementedError(f"{conformance_class} not supported")
-
-    def conforms_to(self, conformance_class: ConformanceClasses) -> bool:
-        """Whether the API conforms to the given standard. This method only checks
-        against the ``"conformsTo"`` property from the API landing page and does not
-        make any additional calls to a ``/conformance`` endpoint even if the API
-        provides such an endpoint.
-
-        Args:
-            conformance_class : The ``ConformanceClasses`` key to check conformance
-                against.
-
-        Return:
-            bool: Indicates if the API conforms to the given spec or URI.
-        """
-
-        # Conformance of None means ignore all conformance as opposed to an
-        #  empty array which would indicate the API conforms to nothing
-        if self._conformance is None:
-            return True
-
-        class_regex = CONFORMANCE_URIS.get(conformance_class.name, None)
-
-        if class_regex is None:
-            raise Exception(f"Invalid conformance class {conformance_class}")
-
-        pattern = re.compile(class_regex)
-
-        if not any(re.match(pattern, uri) for uri in self._conformance):
-            return False
-
-        return True
-
-    def set_conformance(self, conformance: Optional[List[str]]) -> None:
-        """Sets (or clears) the conformance classes for this StacIO."""
-        self._conformance = conformance
 
 
 def _is_url(href: str) -> bool: